+++ console/src/components/Services/Events/EventTriggers/TriggerContainer.tsx
@@ -1,9 +1,13 @@
 import React, { useEffect } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
-
-import { getEventTriggerByName } from '../../../../metadata/selector';
+import { setDriver } from '@/dataSources';
+import {
+  getDataSources,
+  getEventTriggerByName,
+} from '../../../../metadata/selector';
 import { ReduxState } from '../../../../types';
 import { UPDATE_CURRENT_DATA_SOURCE } from '../../Data/DataActions';
+import { getSourceDriver } from '../../Data/utils';
 import { RouterTriggerProps } from '../types';
 
 const TriggerContainer: React.FC<Props> = ({
@@ -11,6 +15,7 @@ const TriggerContainer: React.FC<Props> = ({
   currentTrigger,
   dispatch,
   currentDataSource,
+  driver,
 }) => {
   useEffect(() => {
     if (currentTrigger && currentTrigger.source !== currentDataSource) {
@@ -18,6 +23,7 @@ const TriggerContainer: React.FC<Props> = ({
         type: UPDATE_CURRENT_DATA_SOURCE,
         source: currentTrigger?.source,
       });
+      setDriver(driver);
     }
   }, [currentDataSource, currentTrigger, dispatch]);
 
@@ -29,10 +35,12 @@ const TriggerContainer: React.FC<Props> = ({
 const mapStateToProps = (state: ReduxState, ownProps: RouterTriggerProps) => {
   const triggerName = ownProps.params.triggerName;
   const currentTrigger = getEventTriggerByName(state)(triggerName);
+  const driver = getSourceDriver(getDataSources(state), currentTrigger?.source);
 
   return {
     currentTrigger,
     currentDataSource: state.tables.currentDataSource,
+    driver,
   };
 };