+++ console/src/components/Services/Events/EventTriggers/state.ts
@@ -5,16 +5,15 @@ import {
   ETOperationColumn,
   EventTrigger,
   RetryConf,
+  DatabaseInfo,
 } from '../types';
 import { Header, defaultHeader } from '../../../Common/Headers/Headers';
 import {
   parseServerWebhook,
   parseEventTriggerOperations,
   getETOperationColumns,
-  findETTable,
 } from '../utils';
 import { parseServerHeaders } from '../../../Common/Headers/utils';
-import { Table } from '../../../../dataSources/types';
 import { generateTableDef } from '../../../../dataSources';
 import { QualifiedTable } from '../../../../metadata/types';
 
@@ -59,8 +58,7 @@ const defaultState: LocalEventTriggerState = {
 
 export const parseServerETDefinition = (
   eventTrigger?: EventTrigger,
-  table?: Table,
-  source?: string
+  databaseInfo?: DatabaseInfo
 ): LocalEventTriggerState => {
   if (!eventTrigger) {
     return defaultState;
@@ -73,16 +71,18 @@ export const parseServerETDefinition = (
     eventTrigger.table_name,
     eventTrigger.schema_name
   );
+  const columnInfo =
+    databaseInfo?.[eventTrigger.schema_name]?.[eventTrigger.table_name] ?? [];
 
   return {
     name: eventTrigger.name,
-    source: source ?? '',
+    source: eventTrigger.source ?? '',
     table: etTableDef,
     operations: parseEventTriggerOperations(etDef),
-    operationColumns: table
+    operationColumns: databaseInfo
       ? getETOperationColumns(
           etDef.update ? etDef.update.columns : [],
-          table.columns
+          columnInfo
         )
       : [],
     webhook: parseServerWebhook(etConf.webhook, etConf.webhook_from_env),
@@ -121,6 +121,11 @@ export const useEventTrigger = (initState?: LocalEventTriggerState) => {
             newTableDef = {
               ...newTableDef,
               name: tableName,
+            };
+          } else if (!tableName && !schemaName) {
+            newTableDef = {
+              name: '',
+              schema: '',
             };
           }
           return {
@@ -176,20 +181,12 @@ export const useEventTrigger = (initState?: LocalEventTriggerState) => {
 
 export const useEventTriggerModify = (
   eventTrigger: EventTrigger,
-  allTables: Table[],
-  source: string
+  databaseInfo: DatabaseInfo
 ) => {
-  const table = findETTable(eventTrigger, allTables);
   const { state, setState } = useEventTrigger(
-    parseServerETDefinition(eventTrigger, table, source)
+    parseServerETDefinition(eventTrigger, databaseInfo)
   );
 
-  React.useEffect(() => {
-    if (allTables.length) {
-      const etTable = findETTable(eventTrigger, allTables);
-      setState.bulk(parseServerETDefinition(eventTrigger, etTable, source));
-    }
-  }, [allTables]);
   return {
     state,
     setState,