+++ console/src/dataSources/services/mssql/index.tsx
@@ -1,3 +1,4 @@
+import { TriggerOperation } from '@/components/Common/FilterQuery/state';
 import React from 'react';
 import { DeepRequired } from 'ts-essentials';
 import { DataSourcesAPI } from '../..';
@@ -204,7 +205,7 @@ export const supportedFeatures: DeepRequired<SupportedFeaturesType> = {
   events: {
     triggers: {
       enabled: true,
-      add: false,
+      add: true,
     },
   },
   actions: {
@@ -897,10 +898,44 @@ INNER JOIN sys.schemas sch2
   deleteFunctionSql: () => {
     return '';
   },
-  getEventInvocationInfoByIDSql: () => {
-    return '';
+  // temporary workaround SQL query (till we get an API) as ODBC server library does not support some data types
+  // https://github.com/hasura/graphql-engine-mono/issues/4641
+  getEventInvocationInfoByIDSql: (
+    logTableDef: QualifiedTable,
+    eventLogTable: QualifiedTable,
+    eventId: string
+  ) => {
+    const sql = `SELECT CONVERT(varchar(MAX), original_table.id) AS "id", CONVERT(varchar(MAX), original_table.event_id) AS "event_id",  
+        original_table.status,  CONVERT(varchar(MAX), original_table.request) AS "request", CONVERT(varchar(MAX), original_table.response) AS "response", 
+        CONVERT(varchar(MAX), CAST(original_table.created_at as datetime2)) AS "created_at", CONVERT(varchar(MAX), data_table.id) AS "id", data_table.schema_name, 
+        data_table.table_name, data_table.trigger_name, CONVERT(varchar(MAX), data_table.payload) AS "payload", data_table.delivered, data_table.error,
+        data_table.tries, CONVERT(varchar(MAX), CAST(data_table.created_at as datetime2)) AS "created_at", CONVERT(varchar(MAX), data_table.locked) AS "locked", 
+        CONVERT(varchar(MAX), data_table.next_retry_at) AS "next_retry_at", data_table.archived 
+        FROM "${logTableDef.schema}"."${logTableDef.name}" AS original_table JOIN "${eventLogTable.schema}"."${eventLogTable.name}" 
+        AS data_table ON original_table.event_id = data_table.id
+        WHERE original_table.event_id = '${eventId}'
+        ORDER BY original_table.created_at DESC `;
+    return sql;
   },
-  getDatabaseInfo: '',
+  getDatabaseInfo: `
+SELECT
+  TABLE_SCHEMA AS table_schema,
+  TABLE_NAME AS table_name,
+  COLUMN_NAME AS "column.columnName",
+  DATA_TYPE AS "column.columnType"
+FROM INFORMATION_SCHEMA.COLUMNS
+WHERE
+    TABLE_SCHEMA NOT IN (
+      'guest', 'INFORMATION_SCHEMA', 'sys',
+      'db_owner', 'db_securityadmin', 'db_accessadmin',
+      'db_backupoperator', 'db_ddladmin', 'db_datawriter',
+      'db_datareader', 'db_denydatawriter', 'db_denydatareader'
+    )
+ORDER BY
+  table_schema,
+  table_name
+FOR JSON path;
+`,
   getTableInfo: (tables: QualifiedTable[]) => `
 SELECT
 	o.name AS table_name,
@@ -970,5 +1005,121 @@ WHERE
   },
   getAlterFunctionCommentSql: () => {
     return '';
+  },
+  // temporary workaround SQL query (till we get an API) as ODBC server library does not support some data types
+  // https://github.com/hasura/graphql-engine-mono/issues/4641
+  getDataTriggerLogsCountQuery: (
+    triggerName: string,
+    triggerOp: TriggerOperation
+  ): string => {
+    const triggerTypes = {
+      pending: 'pending',
+      processed: 'processed',
+      invocation: 'invocation',
+    };
+    const eventRelTable = `"hdb_catalog"."event_log"`;
+    const eventInvTable = `"hdb_catalog"."event_invocation_logs"`;
+
+    let logsCountQuery = `SELECT
+    COUNT(*)
+    FROM ${eventRelTable} data_table
+    WHERE data_table.trigger_name = '${triggerName}' `;
+
+    switch (triggerOp) {
+      case triggerTypes.pending:
+        logsCountQuery += `AND delivered=0 AND error=0 AND archived=0;`;
+        break;
+
+      case triggerTypes.processed:
+        logsCountQuery += `AND (delivered=1 OR error=1) AND archived=0;`;
+        break;
+      case triggerTypes.invocation:
+        logsCountQuery = `SELECT
+          COUNT(*)
+          FROM ${eventInvTable} original_table JOIN ${eventRelTable} data_table
+          ON original_table.event_id = data_table.id
+          WHERE data_table.trigger_name = '${triggerName}' `;
+        break;
+      default:
+        break;
+    }
+    return logsCountQuery;
+  },
+  // temporary workaround SQL query (till we get an API) as ODBC server library does not support some data types
+  // https://github.com/hasura/graphql-engine-mono/issues/4641
+  getDataTriggerLogsQuery: (
+    triggerOp: TriggerOperation,
+    triggerName: string,
+    limit?: number,
+    offset?: number
+  ): string => {
+    const triggerTypes = {
+      pending: 'pending',
+      processed: 'processed',
+      invocation: 'invocation',
+    };
+    const eventRelTable = `"hdb_catalog"."event_log"`;
+    const eventInvTable = `"hdb_catalog"."event_invocation_logs"`;
+    let sql = '';
+
+    switch (triggerOp) {
+      case triggerTypes.pending:
+        sql = `SELECT CONVERT(varchar(MAX), id) AS "id", schema_name, table_name, trigger_name, 
+        CONVERT(varchar(MAX), payload) AS "payload", delivered, error, tries, CONVERT(varchar(MAX), CAST(created_at as datetime2) ) AS "created_at",
+        CONVERT(varchar(MAX), locked) AS "locked", CONVERT(varchar(MAX), next_retry_at) AS "next_retry_at", archived 
+        FROM ${eventRelTable} AS data_table
+        WHERE data_table.trigger_name = '${triggerName}'  
+        AND delivered=0 AND error=0 AND archived=0 ORDER BY created_at DESC `;
+        break;
+
+      case triggerTypes.processed:
+        sql = `SELECT CONVERT(varchar(MAX), id) AS "id", schema_name, table_name, trigger_name, 
+        CONVERT(varchar(MAX), payload) AS "payload", delivered, error, tries, CONVERT(varchar(MAX), CAST(created_at as datetime2) ) AS "created_at",
+        CONVERT(varchar(MAX), locked) AS "locked", CONVERT(varchar(MAX), next_retry_at) AS "next_retry_at", archived 
+        FROM ${eventRelTable} AS data_table 
+        WHERE data_table.trigger_name = '${triggerName}' 
+        AND (delivered=1 OR error=1) AND archived=0 ORDER BY created_at DESC `;
+        break;
+
+      case triggerTypes.invocation:
+        sql = `SELECT CONVERT(varchar(MAX), original_table.id) AS "id", CONVERT(varchar(MAX), original_table.event_id) AS "event_id",  
+        original_table.status,  CONVERT(varchar(MAX), original_table.request) AS "request", CONVERT(varchar(MAX), original_table.response) AS "response", 
+        CONVERT(varchar(MAX), CAST(original_table.created_at as datetime2)) AS "created_at", CONVERT(varchar(MAX), data_table.id) AS "id", data_table.schema_name, 
+        data_table.table_name, data_table.trigger_name, CONVERT(varchar(MAX), data_table.payload) AS "payload", data_table.delivered, data_table.error,
+        data_table.tries, CONVERT(varchar(MAX), CAST(data_table.created_at as datetime2)) AS "created_at", CONVERT(varchar(MAX), data_table.locked) AS "locked", 
+        CONVERT(varchar(MAX), data_table.next_retry_at) AS "next_retry_at", data_table.archived 
+        FROM ${eventInvTable} AS original_table JOIN ${eventRelTable} AS data_table ON original_table.event_id = data_table.id
+        WHERE data_table.trigger_name = '${triggerName}' 
+        ORDER BY original_table.created_at DESC `;
+        break;
+      default:
+        break;
+    }
+
+    if (offset) {
+      sql += ` OFFSET ${offset} ROWS`;
+    } else {
+      sql += ` OFFSET 0 ROWS`;
+    }
+
+    if (limit) {
+      sql += ` FETCH NEXT ${limit} ROWS ONLY;`;
+    } else {
+      sql += ` FETCH NEXT 10 ROWS ONLY;`;
+    }
+
+    return sql;
+  },
+  // temporary workaround SQL query (till we get an API) as ODBC server library does not support some data types
+  // https://github.com/hasura/graphql-engine-mono/issues/4641
+  getDataTriggerInvocations: (eventId: string): string => {
+    const eventInvTable = `"hdb_catalog"."event_invocation_logs"`;
+    const sql = `SELECT CONVERT(varchar(MAX), id) AS "id", CONVERT(varchar(MAX), event_id) AS "event_id",  
+    status,  CONVERT(varchar(MAX), request) AS "request", CONVERT(varchar(MAX), response) AS "response", 
+    CONVERT(varchar(MAX), CAST(created_at as datetime2)) AS "created_at"
+    FROM ${eventInvTable} 
+    WHERE event_id = '${eventId}' 
+    ORDER BY created_at DESC;`;
+    return sql;
   },
 };